Add host-side events
Adds the host-side commands that deal with events. This does not include the
vkCmd* that also handle events.

I think using set(&mut self) is correct, even though it does not directly
mutate the structure. It means that Rust will deal with the host-side
synchronization at compile time.

TODO: Revisit for Fences and other structures.